Add reload progress bar to the flight overlay
