+++ Source/UI/Components/ReloadProgressBar/ReloadProgressBar.tscn
@@ -1,0 +1,11 @@
+[gd_scene load_steps=2 format=3 uid="uid://c63mfgyxtk0av"]
+
+[ext_resource type="Script" uid="uid://c5hwukmsf4tck" path="res://Source/UI/Components/ReloadProgressBar/ReloadProgressBar.cs" id="1_dukq8"]
+
+[node name="ReloadProgressBar" type="ProgressBar"]
+custom_minimum_size = Vector2(0, 70)
+size_flags_vertical = 3
+value = 50.0
+fill_mode = 1
+show_percentage = false
+script = ExtResource("1_dukq8")
