+++ Source/UI/Overlays/AimingOverlay/AimingOverlay.tscn
@@ -1,4 +1,4 @@
-[gd_scene load_steps=14 format=3 uid="uid://bxtkrlhh7lgbf"]
+[gd_scene load_steps=15 format=3 uid="uid://bxtkrlhh7lgbf"]
 
 [ext_resource type="Theme" uid="uid://b4g3uaa2tkx1o" path="res://Assets/Themes/UITheme.tres" id="1_ahla6"]
 [ext_resource type="Script" uid="uid://vbyo8ke0rpog" path="res://Source/UI/Components/Buttons/UIButton/UiButton.cs" id="1_it1uj"]
@@ -13,8 +13,9 @@
 [ext_resource type="Texture2D" uid="uid://xggg224p4r8j" path="res://Assets/UI/Buttons/Square-Medium/Pause/Hover.svg" id="4_vdcg2"]
 [ext_resource type="Texture2D" uid="uid://cfomq2sd4rw8x" path="res://Assets/UI/Buttons/Square-Medium/ArrowDown/Default.png" id="9_uqvhh"]
 [ext_resource type="Texture2D" uid="uid://dtnuag5l66wm8" path="res://Assets/UI/Buttons/Square-Medium/ArrowDown/Hover.png" id="10_y4jlo"]
+[ext_resource type="PackedScene" uid="uid://c63mfgyxtk0av" path="res://Source/UI/Components/ReloadProgressBar/ReloadProgressBar.tscn" id="14_vdcg2"]
 
-[node name="AimingOverlay" type="Control" node_paths=PackedStringArray("RoundNumberLabel", "EnemiesRemainingLabel", "AmmoLabel", "ReloadProgressBar", "ReloadButton", "ShootButton", "PauseButton", "RotateLeftButton", "RotateRightButton", "RotateDownButton", "RotateUpButton")]
+[node name="AimingOverlay" type="Control" node_paths=PackedStringArray("RoundNumberLabel", "EnemiesRemainingLabel", "AmmoLabel", "ShootButton", "PauseButton", "RotateLeftButton", "RotateRightButton", "RotateDownButton", "RotateUpButton")]
 layout_mode = 3
 anchors_preset = 15
 anchor_right = 1.0
@@ -26,8 +27,6 @@ script = ExtResource("2_cw5iy")
 RoundNumberLabel = NodePath("MarginContainer/VBoxContainer/HBoxContainer/HBoxContainer/RoundNumberLabel")
 EnemiesRemainingLabel = NodePath("MarginContainer/VBoxContainer/HBoxContainer/EnemiesRemaining")
 AmmoLabel = NodePath("MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer/AmmoLabel")
-ReloadProgressBar = NodePath("MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/ReloadProgressBar")
-ReloadButton = NodePath("MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer/ReloadButton")
 ShootButton = NodePath("MarginContainer/VBoxContainer/HBoxContainer2/MarginContainer/VBoxContainer/ShootButton")
 PauseButton = NodePath("MarginContainer/VBoxContainer/HBoxContainer/HBoxContainer/PauseButton")
 RotateLeftButton = NodePath("MarginContainer/VBoxContainer/HBoxContainer2/RotateLeftButton")
@@ -172,27 +171,14 @@ size_flags_horizontal = 3
 layout_mode = 2
 size_flags_vertical = 3
 
-[node name="ReloadButton" type="Button" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer"]
-layout_mode = 2
-size_flags_horizontal = 6
-size_flags_vertical = 4
-text = "Reload"
-script = ExtResource("1_it1uj")
-metadata/_custom_type_script = "uid://vbyo8ke0rpog"
-
 [node name="AmmoLabel" type="Label" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer"]
 layout_mode = 2
 size_flags_horizontal = 10
 text = "3/5"
 horizontal_alignment = 2
 
-[node name="ReloadProgressBar" type="ProgressBar" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2"]
-custom_minimum_size = Vector2(0, 70)
+[node name="ReloadProgressBar" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2" instance=ExtResource("14_vdcg2")]
 layout_mode = 2
-size_flags_vertical = 3
-value = 50.0
-fill_mode = 1
-show_percentage = false
 
 [connection signal="pressed" from="MarginContainer/VBoxContainer/HBoxContainer/HBoxContainer/PauseButton" to="." method="OnPauseButtonPressed"]
 [connection signal="button_down" from="MarginContainer/VBoxContainer/HBoxContainer/MarginContainer/RotateUpButton" to="." method="OnRotateUpButtonDown"]
@@ -204,4 +190,3 @@ show_percentage = false
 [connection signal="button_up" from="MarginContainer/VBoxContainer/HBoxContainer2/RotateRightButton" to="." method="OnRotateRightButtonUp"]
 [connection signal="button_down" from="MarginContainer/VBoxContainer/HBoxContainer3/MarginContainer/RotateDownButton" to="." method="OnRotateDownButtonDown"]
 [connection signal="button_up" from="MarginContainer/VBoxContainer/HBoxContainer3/MarginContainer/RotateDownButton" to="." method="OnRotateDownButtonUp"]
-[connection signal="pressed" from="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer/ReloadButton" to="." method="OnReloadButtonPressed"]
