+++ Source/UI/Overlays/FlightOverlay/FlightOverlay.tscn
@@ -1,9 +1,10 @@
-[gd_scene load_steps=5 format=3 uid="uid://dgii5mcu5tqm0"]
+[gd_scene load_steps=6 format=3 uid="uid://dgii5mcu5tqm0"]
 
 [ext_resource type="Theme" uid="uid://b4g3uaa2tkx1o" path="res://Assets/Themes/UITheme.tres" id="1_wccxp"]
 [ext_resource type="Script" uid="uid://5v6de3h84uym" path="res://Source/UI/Overlays/FlightOverlay/FlightOverlay.cs" id="2_1gpin"]
 [ext_resource type="Texture2D" uid="uid://ywo4def610hj" path="res://Assets/UI/Buttons/Square-Medium/Pause/Default.svg" id="2_4bla2"]
 [ext_resource type="Texture2D" uid="uid://xggg224p4r8j" path="res://Assets/UI/Buttons/Square-Medium/Pause/Hover.svg" id="3_vacut"]
+[ext_resource type="PackedScene" uid="uid://c63mfgyxtk0av" path="res://Source/UI/Components/ReloadProgressBar/ReloadProgressBar.tscn" id="5_xapji"]
 
 [node name="FlightOverlay" type="Control" node_paths=PackedStringArray("PauseButton")]
 layout_mode = 3
@@ -71,6 +72,31 @@ layout_mode = 2
 size_flags_horizontal = 3
 
 [node name="HBoxContainer" type="HBoxContainer" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2"]
+layout_mode = 2
+
+[node name="VBoxContainer" type="VBoxContainer" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer"]
+layout_mode = 2
+size_flags_horizontal = 3
+
+[node name="VBoxContainer3" type="VBoxContainer" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer"]
+layout_mode = 2
+size_flags_horizontal = 3
+
+[node name="VBoxContainer2" type="VBoxContainer" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer"]
+layout_mode = 2
+size_flags_horizontal = 3
+
+[node name="HBoxContainer" type="HBoxContainer" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer/VBoxContainer2"]
+layout_mode = 2
+size_flags_vertical = 3
+
+[node name="AmmoLabel" type="Label" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer/VBoxContainer2/HBoxContainer"]
+layout_mode = 2
+size_flags_horizontal = 10
+text = "3/5"
+horizontal_alignment = 2
+
+[node name="ReloadProgressBar" parent="MarginContainer/VBoxContainer/HBoxContainer3/VBoxContainer2/HBoxContainer/VBoxContainer2" instance=ExtResource("5_xapji")]
 layout_mode = 2
 
 [connection signal="pressed" from="MarginContainer/VBoxContainer/HBoxContainer/HBoxContainer/PauseButton" to="." method="OnPauseButtonPressed"]
